⚡ :: (#26) 코드수정
- 파일 위치 변경
- menu hover 삭제
